refactor module name convention
